Split config between ZZ and ZH in a nice way (only 2018 for now)
